Add a jpeg-loss experiment script
The experiment script generates random bits, and attempts to write them to pixels using one of a few different encoding strategies. The pixels are jpeg-encoded and then an attempt is made to recover the original bits.

The BlackAndWhiteMapper is the only mapper in the file to successfully preserve the information in question.